7: Fix Error-Codes consistency r=Jaskaranbir a=Jaskaranbir

+++ model/error.go
@@ -2,12 +2,12 @@ package model
 
 // Error describes any error.
 type Error struct {
-	Code    int    `json:"code,omitempty"`
+	Code    int16  `json:"code,omitempty"`
 	Message string `json:"message,omitempty"`
 }
 
 // NewError is a convenient awy to create an Error-instance.
-func NewError(code int, msg string) *Error {
+func NewError(code int16, msg string) *Error {
 	return &Error{
 		Code:    code,
 		Message: msg,
